requirements.txt
requests
requests_oauthlib
mirri_utils